+++ resources/views/discover/home.blade.php
@@ -8,6 +8,5 @@
 
 @push('scripts')
 <script type="text/javascript" src="{{ mix('js/discover.js') }}"></script>
-<script type="text/javascript" src="{{ mix('js/compose.js') }}"></script>
 <script type="text/javascript">App.boot();</script>
 @endpush